requirements.txt
fastapi
uvicorn
pillow
python-dotenv
transformers
torch
requests
accelerate